+++ x/superfluid/keeper/gov/gov_test.go
@@ -61,8 +61,9 @@ func (suite *KeeperTestSuite) TestHandleSetSuperfluidAssetsProposal() {
 		expectErr      bool
 	}
 	testCases := []struct {
-		name    string
-		actions []Action
+		name          string
+		actions       []Action
+		expectedEvent []string
 	}{
 		{
 			"happy path flow",
@@ -74,6 +75,7 @@ func (suite *KeeperTestSuite) TestHandleSetSuperfluidAssetsProposal() {
 					false, []types.SuperfluidAsset{asset1}, []types.SuperfluidAsset{}, false,
 				},
 			},
+			[]string{types.TypeEvtSetSuperfluidAsset, types.TypeEvtRemoveSuperfluidAsset},
 		},
 		{
 			"token does not exist",
@@ -85,6 +87,7 @@ func (suite *KeeperTestSuite) TestHandleSetSuperfluidAssetsProposal() {
 					false, []types.SuperfluidAsset{asset2}, []types.SuperfluidAsset{asset1}, true,
 				},
 			},
+			[]string{types.TypeEvtSetSuperfluidAsset, types.TypeEvtRemoveSuperfluidAsset},
 		},
 	}
 
@@ -132,6 +135,7 @@ func (suite *KeeperTestSuite) TestHandleSetSuperfluidAssetsProposal() {
 					suite.Require().Error(err)
 				} else {
 					suite.Require().NoError(err)
+					suite.AssertEventEmitted(suite.ctx, tc.expectedEvent[i], 1)
 				}
 
 				// check assets individually